+++ laboratory-task-13/src/class/matrix/matrix.hpp
@@ -1,16 +1,17 @@
 #ifndef MATRIX_HPP
 #define MATRIX_HPP
 #include <iostream>
+#include <cstdint>
 
 class Matrix
 {
 private:
-	uint8_t row; // строка
-	uint8_t col; // столбец
+	uint16_t row; // строка
+	uint16_t col; // столбец
 	double** mtr; // двумерный массив
 public:
 	Matrix(); // по умолчанию 
-	Matrix(const uint8_t&, const uint8_t&); // с параметрами
+	Matrix(const uint16_t&, const uint16_t&); // с параметрами
 	Matrix(const Matrix&); // копирования
 	Matrix(Matrix&&) noexcept; // перемещения
 	~Matrix(); // деструктор
